mami/requirements.txt
asgiref==3.7.2
dependency-injector==4.41.0
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
phonenumbers==8.13.24
Pillow==10.1.0
PyJWT==2.8.0
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
